// src/consts.d.ts
export declare enum Orientation {
    V = 1,
    VERTICAL = 1,
    H = 2,
    HORIZONTAL = 2,
}
export declare enum Direction {
    W = 1,
    WEST = 1,
    E = 2,
    EAST = 2,
    N = 3,
    NORTH = 3,
    S = 4,
    SOUTH = 4,
    L = 1,
    LEFT = 1,
    R = 2,
    RIGHT = 2,
    T = 3,
    TOP = 3,
    B = 4,
    BOTTOM = 4,
}
export declare enum AlignH {
    L = 1,
    LEFT = 1,
    C = 2,
    CENTER = 2,
    R = 3,
    RIGHT = 3,
}
export declare enum AlignV {
    T = 1,
    TOP = 1,
    M = 2,
    MIDDLE = 2,
    B = 3,
    BOTTOM = 3,
}
export declare enum Align {
    L = 1,
    LEFT = 1,
    C = 2,
    CENTER = 2,
    R = 3,
    RIGHT = 3,
    T = 1,
    TOP = 1,
    M = 2,
    MIDDLE = 2,
    B = 3,
    BOTTOM = 3,
}